Don't update games/users which don't need it
The queue can be added to multiple times. Games and users shouldn’t get
ran if they don’t need to.

+++ app/workers/update_user_worker.rb
@@ -4,17 +4,22 @@ class UpdateUserWorker
 
   def perform(user_id)
     user = User.find(user_id)
-    user.games.delete_all # Remove old associations.
 
-    user.collection.each do |item|
-      if item["status"].first["own"] == "1"
-        game = Game.find_or_create_by(:bgg_id => item['objectid'])
-        user.games << game
-        game.update_bgg_data # This will kick off a worker.
+    if user.needs_updating?
+      user.games.delete_all # Remove old associations.
+
+      user.collection.each do |item|
+        if item["status"].first["own"] == "1"
+          game = Game.find_or_create_by(:bgg_id => item['objectid'])
+          user.games << game
+          # game.update_bgg_data # This will kick off a worker.
+        end
       end
-    end
 
-    user.updated_at = Time.now
-    user.save
+      user.updated_at = Time.now
+      user.save
+    else
+      logger.info "User #{user.name} has already been updated."
+    end
   end
 end
